Zephyr Support

+++ src/AudioTools/AudioLibs/AudioBoardStream.h
@@ -24,8 +24,8 @@ class AudioBoardStream : public I2SCodecStream {
     }
     AudioDriverKey key;
     AudioBoard *p_board;
-    int id() override { return key | 0x400; }
-    bool readValue() override { return p_board->isKeyPressed(key); }
+    int id() override { return (int) key | 0x400; }
+    bool readValue() override { return p_board->isKeyPressed((int)key); }
   };
 
  public:
